pyproject.toml
[project]
name = "cfbot"
version = "0.1.0"
description = "Telegram bot for managing Cloudflare DNS records"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "aiohttp-socks>=0.11.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
